last stable version

+++ include/RectObstacle.h
@@ -9,7 +9,7 @@ public:
     void updateFromFluid(FluidGrid& grid, float dt) override;
 
 private:
-    int m_x, m_y, m_w, m_h;
+    float m_x, m_y, m_w, m_h;
     float m_vx = 0.f, m_vy = 0.f;
     float m_xf = 0.f, m_yf = 0.f;
     int m_gridN; // To know grid dimensions for drawing
